+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 ### Added
 
   - Profile - Enable or disable 2FA (Two factor authentication)
+  - NuvlaBox - new options for opting in on host-level management, at NuvlaBox creation time
+  - NuvlaBox - new tab for managing playbooks
+  - NuvlaBox - new buttons for playbook management operations
 
 ### Changed